Moving logic to the abstract Validator
